requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
Django==3.2.5
idna==2.10
image==1.5.33
mss==6.1.0
numpy==1.21.0
opencv-python==4.5.2.54
Pillow==8.3.1
ppad==0.0.11
progressbar2==3.53.1
pure-python-adb==0.3.0.dev0
python-dateutil==2.8.1
python-utils==2.5.6
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
